LayoutTests/inspector/canvas/setRecordingAutoCaptureFrameCount-expected.txt

Test that contexts created after calling Canvas.setRecordingAutoCaptureFrameCount are properly recorded.


== Running test suite: Canvas.setRecordingAutoCaptureFrameCount
-- Running test case: Canvas.setRecordingAutoCaptureFrameCount.2D.None
PASS: Canvas should have no finished recordings.

-- Running test case: Canvas.setRecordingAutoCaptureFrameCount.2D.Single
PASS: Recording started.
PASS: Recording stopped.
PASS: Recording should have one frame.
PASS: Recording should have one action.
PASS: Canvas should have one finished recording.

-- Running test case: Canvas.setRecordingAutoCaptureFrameCount.2D.Multiple
PASS: Recording started.
PASS: Canvas should have no finished recordings.
PASS: Canvas should be actively recording.
PASS: Recording should have 2 frames.

-- Running test case: Canvas.setRecordingAutoCaptureFrameCount.WebGL.None
PASS: Canvas should have no finished recordings.

-- Running test case: Canvas.setRecordingAutoCaptureFrameCount.WebGL.Single
PASS: Recording started.
PASS: Recording stopped.
PASS: Recording should have one frame.
PASS: Recording should have one action.
PASS: Canvas should have one finished recording.

-- Running test case: Canvas.setRecordingAutoCaptureFrameCount.WebGL.Multiple
PASS: Recording started.
PASS: Canvas should have no finished recordings.
PASS: Canvas should be actively recording.
PASS: Recording should have 2 frames.
